Break share main file into subcommands files - upload, download, list
